fix(tests): avoid use of invalid escape sequence
When using `-Werror` this results in the `DeprecationWarning` for
invalid escape sequences becoming a `SyntaxError`. At some point this
will become a `SyntaxError` by default in a future version of Python.

The `match` argument of `pytest.raises()` is expecting a regular
expression, so it makes sense to always use a raw string.